# internal/api/v1/docs/authAPI_swagger.yaml
basePath: /v1
definitions:
...
  models.DeviceCallbackResponse:
    properties:
      identifier:
        type: string
    type: object
  models.DeviceFlowResponse:
    properties:
      device_code:
        type: string
      polling_rate:
        type: integer
      user_code:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@loopholelabs.io
    name: API Support
  description: Auth API, V1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://loopholelabs.io/privacy
  title: Auth API V1
  version: "1.0"
paths:
...
  /device/callback:
    post:
      consumes:
      - application/json
      description: DeviceCallback validates the device code and returns the flow identifier
      parameters:
      - description: device code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceCallbackResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: DeviceCallback validates the device code and returns the flow identifier
      tags:
      - device
      - callback
  /device/flow:
    post:
      consumes:
      - application/json
      description: DeviceFlow starts the device code flow
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceFlowResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: DeviceFlow starts the device code flow
      tags:
      - device
      - login
  /device/poll:
    post:
      consumes:
      - application/json
      description: DevicePoll polls the device code flow using the user code
      parameters:
      - description: user code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: DevicePoll polls the device code flow using the user code
      tags:
      - device
      - poll
  /github/callback:
    get:
      consumes:
      - application/json
      description: GithubCallback logs in a user with Github
      produces:
      - application/json
      responses:
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: Redirects to Next URL
              type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GithubCallback logs in a user with Github
      tags:
      - github
      - callback
  /github/login:
    get:
      consumes:
      - application/json
      description: GithubLogin logs in a user with Github
      parameters:
      - description: Next Redirect URL
        in: query
        name: next
        type: string
      - description: Organization
        in: query
        name: organization
        type: string
      - description: Device Flow Identifier
        in: query
        name: identifier
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: Redirects to Github
              type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GithubLogin logs in a user with Github
      tags:
      - github
      - login
  /google/callback:
    get:
      consumes:
      - application/json
      description: GoogleCallback logs in a user with Google
      produces:
      - application/json
      responses:
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: Redirects to Next URL
              type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GoogleCallback logs in a user with Google
      tags:
      - google
      - callback
  /google/login:
    get:
      consumes:
      - application/json
      description: GoogleLogin logs in a user with Google
      parameters:
      - description: Next Redirect URL
        in: query
        name: next
        type: string
      - description: Organization
        in: query
        name: organization
        type: string
      - description: Device Flow Identifier
        in: query
        name: identifier
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: Redirects to Google
              type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GoogleLogin logs in a user with Google
      tags:
      - google
      - login
...
schemes:
- https
swagger: "2.0"